mobile: split MyTeamScreen.tsx into focused sibling files
586L → 5 files in mobile/src/screens/MyTeam/:
- index.tsx (274L) — MyTeamScreen + MyTeamContent + ControlBar +
  ControlButton + NoTeamIdView + toMyTeamRow. Keeps the orchestration
  + small leaf states together.
- Header.tsx (87L) — Header + PicksUnavailableNote + ChipBanner.
  The "above the table" banner-style components.
- NameCell.tsx (46L) — MyTeamNameCell. The pinned-name-column row
  renderer passed to PlayerListTable.
- types.ts (13L) — MyTeamRow type, shared between index and NameCell.
- styles.ts (170L) — shared makeStyles block.

Lighter touch than Transfers because MyTeamScreen was half the size
and its sub-components were less numerous; the orchestrator + small
ControlButton/NoTeamIdView leaves stay co-located rather than moved
out for their own sake.
